Fix plot missing time info #16
Plot function was sometimes messing with time info.
This is caused by the times_inf() functions which add desat times
and remove duplicated ones. Then if desat time is the same as the
depth time, only one is shown.
Also options where either time_print or deco_print. Now both can be
added to the plot. If both are present, the text is bold but this is a
minor issue.

Plot check still require human check with the dev/graphic_vignette,
which is difficult and explain why all the graphic script is not tested.

Resolves: #16

gowachin
maxime.jaunatre@yahoo.fr
24/03/2022